% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2load.R
\name{abf2.load}
\alias{abf2.load}
\title{Load abf2 data file.}
\usage{
abf2.load(filename)
}
\arguments{
\item{filename}{}
}
\value{
Loaded abf2 data
}
\description{
Load abf2 data file.
}
\examples{
data <- abf2.load("data/example.abf")
}
