% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2load.R
\name{abf2.load_in_folder}
\alias{abf2.load_in_folder}
\title{Load abf2 data files in a folder}
\usage{
abf2.load_in_folder(folder, filename_list)
}
\arguments{
\item{folder}{}

\item{filename_list}{}
}
\value{
A list of loaded abf2 data.
}
\description{
Load abf2 data files in a folder
}
